modem: hl7800: use correct timeouts on KTCPSND to avoid internal deadlock
The response of a KTCPSND has two phases. According to documentation by wireless the timeout is 60 seconds.
The fix respects the timeout on the second phase, too (waiting for OK or errors from modem). Previously only the first phase used 60 seconds and the second phase used 5 seconds.

Without this fix the hl7800 will lock the tcp stack for the current socket indefinitely if another socket operation is performed before the response from the modem is received.

Additionaly all timeouts are adjusted to be at least one second longer as the documented timeout from wireless. This avoids races between the hl7800 and the driver.